+++ src/main/java/org/mmarini/wheelly/apis/WheellyContactsMessage.java
@@ -64,7 +64,6 @@ public record WheellyContactsMessage(long localTime, long simulationTime, long r
      * @param line           the status string
      * @param clockConverter the clock converter
      */
-
     public static WheellyContactsMessage create(Timed<String> line, ClockConverter clockConverter) {
         long time = line.time(TimeUnit.MILLISECONDS);
         String[] params = line.value().split(" ");
@@ -78,6 +77,25 @@ public record WheellyContactsMessage(long localTime, long simulationTime, long r
         boolean canMoveForward = Integer.parseInt(params[4]) != 0;
         boolean canMoveBackward = Integer.parseInt(params[5]) != 0;
         long simTime = clockConverter.fromRemote(remoteTime);
+        return new WheellyContactsMessage(time,
+                simTime, remoteTime, frontSensors, rearSensors,
+                canMoveForward, canMoveBackward
+        );
+    }
+
+    public static WheellyContactsMessage create(Timed<String> line, long timeOffset) {
+        long time = line.time(TimeUnit.MILLISECONDS);
+        String[] params = line.value().split(" ");
+        if (params.length != NO_PARAMS) {
+            throw new IllegalArgumentException(format("Wrong contacts message \"%s\" (#params=%d)", line.value(), params.length));
+        }
+        long remoteTime = parseLong(params[1]);
+        boolean frontSensors = Integer.parseInt(params[2]) != 0;
+        boolean rearSensors = Integer.parseInt(params[3]) != 0;
+
+        boolean canMoveForward = Integer.parseInt(params[4]) != 0;
+        boolean canMoveBackward = Integer.parseInt(params[5]) != 0;
+        long simTime = time - timeOffset;
         return new WheellyContactsMessage(time,
                 simTime, remoteTime, frontSensors, rearSensors,
                 canMoveForward, canMoveBackward