+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue [#561](https://github.com/m-marini/wheellyj/issues/561): Add label direction to camera event
+
 ### Fixed
 
 - Issue [#557](https://github.com/m-marini/wheellyj/issues/557): Ghost labels in simulated robot